pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "setuptools_scm[toml]>=7.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "RepRepBuild"
authors = [
  { name="Toon Verstraelen", email="toon.verstraelen@ugent.be" },
]
description = "Build tool for Reproducible Reporting"
readme = "README.md"
license = {file = "COPYING"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Framework :: Matplotlib",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "attrs",
    "beautifulsoup4",
    "bibsane",
    "jinja2",
    "markdown",
    "ninja",
    "numpy",
    "parse",
    "pymupdf",
    "pyyaml",
    "requests",
    "scipy",
    "setuptools",
    "tqdm",
    "watchdog",
]
dynamic = ["version"]

[project.urls]
Issues = "https://github.com/reproducible-reporting/reprepbuild/issues"
Source = "https://github.com/reproducible-reporting/reprepbuild/"
Changelog = "https://github.com/reproducible-reporting/reprepbuild/blob/main/CHANGELOG.md"

[project.scripts]
rr = "reprepbuild.__main__:main"
rrr = "reprepbuild.repeat:main"
rr-check-hrefs = "reprepbuild.scripts.check_hrefs:main"
rr-generator = "reprepbuild.scripts.generator:main"
rr-latex-flat = "reprepbuild.scripts.latex_flat:main"
rr-latex-zip = "reprepbuild.scripts.latex_zip:main"
rr-manifest = "reprepbuild.scripts.manifest:main"
rr-pdf-add-notes = "reprepbuild.scripts.pdf_add_notes:main"
rr-pdf-normalize = "reprepbuild.scripts.pdf_normalize:main"
rr-pdf-nup = "reprepbuild.scripts.pdf_nup:main"
rr-python-script = "reprepbuild.scripts.python_script:main"
rr-render = "reprepbuild.scripts.render:main"
rr-zip = "reprepbuild.scripts.zip:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "UP", "B", "I", "PGH", "PL", "RUF", "C"]
line-length = 100
target-version = "py311"
ignore = ["PLR2004", "PLR0913", "PLR0912", "PLW2901", "PLR0915", "C901"]

[tool.setuptools_scm]
write_to = "src/reprepbuild/_version.py"
version_scheme = "post-release"
local_scheme = "no-local-version"
